Pag 206 end. Beginnign with the improving the categories
