+++ SearchViewController.swift
@@ -20,14 +20,8 @@ class SearchViewController: UIViewController {
     
     // Variables globales/instancia.
     
+    let search = Search()
     var landscapeViewController: LandscapeViewController?
-    
-    var searchResults: [SearchResult] = []
-    // Es opcional, ya que no tendremos un dataTask,
-    // hasta que el usuario haga una búsqueda --> ?.
-    var dataTask: URLSessionDataTask?
-    var hasSearched = false
-    var isLoading = false
     
     // MARK: Structs
     
@@ -71,35 +65,6 @@ class SearchViewController: UIViewController {
         // Dispose of any resources that can be recreated.
     }
     
-    
-    // MARK: NetWork
-    
-    // Función que devuelve un objeto URL válido.
-    func iTunesURL(searchText: String, category: Int) -> URL {
-        
-        let entityName: String
-        
-        switch category {
-        case 1:  entityName = "musicTrack"
-        case 2:  entityName = "software"
-        case 3:  entityName = "ebook"
-        default: entityName = ""
-            
-        }
-        
-        // Escapamos los espacios en  blanco entre los Strings de entrada en
-        // la searchBar, codifica en 'UTF-8' que casi siempre va a funcionar.
-        let escapeSearchText = searchText.addingPercentEncoding(withAllowedCharacters: CharacterSet.urlQueryAllowed)!
-        
-        // Recibimos en el parámetro el texto a buscar que se adjunta al final
-        //  de la url dada y el índice de la categoría (books, software, etc).
-        let urlString = String(format: "https://itunes.apple.com/search?term=%@&limit=200&entity=%@", escapeSearchText, entityName)
-        let url = URL(string: urlString)
-        // Como URL(String) es uno de los inicializadores 'failable',
-        // devuleve un opcional '?',  de ahí que lo desempaquetemos.
-        return url!
-    }
-    
     // MARK: Actions
     
     @IBAction func segmentChanged(_ sender: UISegmentedControl) {
@@ -107,134 +72,7 @@ class SearchViewController: UIViewController {
     }
     
     
-    // MARK: ParseJSON
     
-    // Función que convierte los Strings de búsqueda en un diccionario de objetos JSON y los devuelve.
-    func parse(json data: Data) -> [String : Any]? {
-        
-        do {
-            // En el objeto data ya tenemos el texto JSON.
-            // Los serializamos, covirtiéndolos en Objetos Foundation.
-            // En este caso en un diccionario, de par clave-valor String:Any.
-            return  try JSONSerialization.jsonObject(with: data, options: []) as? [String : Any]
-        } catch  {
-            print("JSON Error: \(error)")
-            return nil
-        }
-    }
-    
-    // Función que hace distinción entre los campos el 'kind' y el 'wrapperType', al igual que lo hace iTunes.
-    // Este método pasa por el diccionario de nivel superior y mira cada resultado de la búsqueda.
-    func parse(dictionary: [String : Any]) -> [SearchResult] {
-        // 1.- Programación defensiva, para asegurarnos que el diccionario tiene
-        // una clave denominada  "results" que contiene un  array en su interior.
-        guard let array = dictionary["results"] as? [Any] else {
-            print("Expected 'results' array'")
-            // Si algo va mal, devolvemos un array vacío.
-            return []
-        }
-        var searchResutls: [SearchResult] = []
-        // 2.- Una vez que se ha cumplido que existe dicho array,
-        // iteramos para obtener cada uno de los elementos del array.
-        for resultDict in array {
-            // 3.- Cada uno de los elementos de array es otro diccionario,
-            // pero como 'resultDict' no es un diccionario como nos gustaría que fuese,
-            // utilizamos 'Any' ya que los elementos del array pueden ser de cualquier tipo.
-            if let resultDict = resultDict as? [String : Any] {
-                // 4.-Para cada uno de los diccionarios, imprimimos el valor de su 'wrapperType'
-                // y el campo 'kind'. La indexación de un diccionario siempre nos da un opcional
-                // por lo que utilizamos 'if --> let' para desenvolver esos valores. Y porque el
-                // diccionario solo contiene valores de tipo Any también lo casteamos a tipo String.
-                var searchResult: SearchResult?
-                if let wrapperType = resultDict["wrapperType"] as? String {
-                    switch wrapperType {
-                    case "track": searchResult = parse(track: resultDict)
-                    case "audiobook": searchResult = parse(audiobook: resultDict)
-                    case "software": searchResult = parse(software: resultDict)
-                    default: break
-                    }
-                } else if let kind = resultDict["kind"] as? String, kind == "ebook" { searchResult = parse(ebook: resultDict) }
-                if let result = searchResult { searchResutls.append(result) }
-            }
-        }
-        // Devolvemos un array con objetos searchResult.
-        return searchResutls
-    }
-    
-    func parse(track dictionary: [String : Any]) -> SearchResult {
-        let searchResult = SearchResult()
-        
-        // Casteamos cada propiedad a su tipo, String y Double en este caso ya que el diccionario es de tipos [String : Any]
-        // por lo que la key siempre será de tipo String, pero el valor puede variar, de ahí el casteo forzado a cada tipo.
-        searchResult.kind = dictionary["kind"] as! String
-        searchResult.name = dictionary["trackName"] as! String
-        searchResult.currency = dictionary["currency"] as! String
-        searchResult.artistName = dictionary["artistName"] as! String
-        searchResult.storeURL = dictionary["trackViewUrl"] as! String
-        searchResult.artworkSmallURL = dictionary["artworkUrl60"] as! String
-        searchResult.artworkLargeURL = dictionary["artworkUrl100"] as! String
-        // Nos aseguramos con if--> let en estos dos parámetros porque
-        // a veces estos datos no vienen en los datos del JSON. Podríamos
-        // asegurar los anteriores, siendo muy precavidos poniéndole el ?
-        // por ejemplo: searchResult.kind = dictionary["kind"] as! String?
-        if let price = dictionary["trakePrice"] as? Double { searchResult.price = price }
-        if let genre = dictionary["primaryGenreName"] as? String { searchResult.genre = genre }
-        
-        return searchResult
-    }
-    
-    func parse(audiobook dictionary: [String: Any]) -> SearchResult {
-        let searchResult = SearchResult()
-        
-        searchResult.name = dictionary["collectionName"] as! String
-        searchResult.artistName = dictionary["artistName"] as! String
-        searchResult.artworkSmallURL = dictionary["artworkUrl60"] as! String
-        searchResult.artworkLargeURL = dictionary["artworkUrl100"] as! String
-        searchResult.storeURL = dictionary["collectionViewUrl"] as! String
-        // No tiene 'kind' los audioBooks, lo hardcodeamos directamente.
-        searchResult.kind = "audiobook"
-        searchResult.currency = dictionary["currency"] as! String
-        
-        if let price = dictionary["collectionPrice"] as? Double { searchResult.price = price }
-        if let genre = dictionary["primaryGenreName"] as? String {searchResult.genre = genre }
-        
-        return searchResult
-    }
-    
-    func parse(software dictionary: [String: Any]) -> SearchResult {
-        let searchResult = SearchResult()
-        
-        searchResult.name = dictionary["trackName"] as! String
-        searchResult.artistName = dictionary["artistName"] as! String
-        searchResult.artworkSmallURL = dictionary["artworkUrl60"] as! String
-        searchResult.artworkLargeURL = dictionary["artworkUrl100"] as! String
-        searchResult.storeURL = dictionary["trackViewUrl"] as! String
-        searchResult.kind = dictionary["kind"] as! String
-        searchResult.currency = dictionary["currency"] as! String
-        
-        if let price = dictionary["price"] as? Double { searchResult.price = price }
-        if let genre = dictionary["primaryGenreName"] as? String { searchResult.genre = genre }
-        return searchResult
-    }
-    
-    func parse(ebook dictionary: [String: Any]) -> SearchResult {
-        let searchResult = SearchResult()
-        
-        searchResult.name = dictionary["trackName"] as! String
-        searchResult.artistName = dictionary["artistName"] as! String
-        searchResult.artworkSmallURL = dictionary["artworkUrl60"] as! String
-        searchResult.artworkLargeURL = dictionary["artworkUrl100"] as! String
-        searchResult.storeURL = dictionary["trackViewUrl"] as! String
-        searchResult.kind = dictionary["kind"] as! String
-        searchResult.currency = dictionary["currency"] as! String
-        
-        if let price = dictionary["price"] as? Double { searchResult.price = price }
-        
-        // Los  audioBooks no tienen un campo  "primaryGenreName", pero sí una gran  variedad de genres.
-        // Utilizamos el método 'joined(separator)' para unir estos nombres de género en un solo String.
-        if let genres: Any = dictionary["genres"] { searchResult.genre = (genres as! [String]).joined(separator: ", ") }
-        return searchResult
-    }
     
     // MARK: UISearchBar
     
@@ -249,7 +87,7 @@ class SearchViewController: UIViewController {
         if segue.identifier == "ShowDetail" {
             let detailViewController = segue.destination as! DetailViewController
             let indexPath = sender as! IndexPath
-            let searchResult = searchResults[indexPath.row]
+            let searchResult = search.searchResults[indexPath.row]
             detailViewController.searchResult = searchResult
         }
     }
@@ -308,7 +146,7 @@ class SearchViewController: UIViewController {
             // El view Controller va a leer del array los resultados en viewDidLoad() para construir el contenido
             // de su scroll view. Pero si accedemos al controller.view antes de establecer los "searchResults",
             // esta propiedad estará todavía  a nil y no podremos hacer nada para rellenar los botones en land.
-            controller.searchResults = searchResults
+            controller.search = search
             
             // 3.- Ajustar el tamaño y la posición del nuevo view controller. Esto hace la vista
             // tan grande como "SearchViewController" que cubre toda la pantalla. El "frame" es el
@@ -410,91 +248,19 @@ extension SearchViewController: UISearchBarDelegate {
     // Función que se llama al pulsar el botón "Buscar/Search" del dispositivo.
     func performSearch() {
         
-        if !searchBar.text!.isEmpty {
-            
-            searchBar.resignFirstResponder()
-            
-            // Cada vez que el usuario pulsa el botón 'search' hacemos primero que la tarea sea cancela por si hubiera alguna búsqueda aún activa.
-            // Gracias al encadenamiento opcional, si alguna búsqueda aún no ha  terminado, 'dataTask' será todavía 'nil'; esto simplemente ignora
-            // la llamada a 'cancel()'. Podíamos haberlo hecho también con if-let. Si ponemos '!' y el opcional es 'nil' se bloqueará la aplicación,
-            // dado que cuando la primera vez que el usuario escribe algo en la searchBar, 'dataTask' aún será 'nil' por lo que se caería la app.
-            dataTask?.cancel()
-            
-            // Activamos el activity indicator y refrescamos la tabla.
-            isLoading = true
-            tableView.reloadData()
-            
-            // Aquí ya se realiza búsqueda, por tanto a true.
-            hasSearched = true
-            searchResults = []
-            
-            // --- *** IMPLEMENTAMOS URLSESSION *** --- //
-            
-            // 1.- Creamos el objeto URL añadiéndole el texto de búsqueda y el índice seleccionado del segment Control.
-            let url = self.iTunesURL(searchText: searchBar.text!, category: segmentControl.selectedSegmentIndex)
-            // 2.- Obtenemos el objeto URLSession, mediante la sesión compartida,
-            // utilizando una configuración predeterminada con respecto al alma-
-            // cenamiento en caché, cookies, y otras cosas web. Podemos crear
-            // nuestra propia configuración, creando nuestros propios objetos
-            // URLSessionConfiguration y URLSession.
-            let session = URLSession.shared
-            // 3.- Crear la tarea  de datos, 'dataTask', para  enviar  solicitudes
-            // HTTPS GET al servidor, pasándole la url, y un closure, el cual será
-            // invocado  cuando la  tarea haya recibido la  respuesta del servidor.
-            dataTask = session.dataTask(with: url, completionHandler: {
-                data, response, error in
-                // 4.- En el interior del closure tenemos tres parámetros, todos opcionales para
-                // que puedan estar a nil, y tienen que desempaquetarse antes de ser utilizados.
-                // Error, contine el error, no conecta con el server, no hay red, o fallo de hard.
-                // Si error es nil, la comunicación con el server fue exitosa. Response, contine
-                // códio y cabeceras de la respuesta del server. Y data, contine los datos reales
-                // que nos envía el server, en este caso en formato JSON.
-                if let error = error as NSError?, error.code == -999 {
-                    print("Failure! \(error)")
-                    return // Cancelamos la búsqueda con 'return', el resto del closure se omite.
-                    
-                    
-                } else if let httpResponse = response as? HTTPURLResponse, httpResponse.statusCode == 200 {
-                    
-                    // Desenvolvemos el objeto data opcional, lo parseamos, y lo
-                    // pasamos  para parsearlo convirtiéndolo  en un diccionario.
-                    if let data = data, let jsonDictionary = self.parse(json: data) {
-                        // Aquí lo convertimos el contenido del diccionario en un objeto serchResults.
-                        self.searchResults = self.parse(dictionary: jsonDictionary)
-                        // Por último lo ordenamos.
-                        self.searchResults.sort(by: <)
-                        
-                        // Actualizamos la tabla con los datos nuevos,
-                        // paramos el spinner. Todo en el hilo principal.
-                        DispatchQueue.main.async {
-                            self.isLoading = false
-                            self.tableView.reloadData()
-                        }
-                        // Salimos.
-                        return
-                        
-                    } // --- *** FIN URLSESSION --- *** //
-                    
-                    // Ponemos este código aquí por si algo ha ido mal, avisando al usuario de que algo ha ido mal.
-                    // Actualizamos la tabla antes, ya que la vista de la tabla necesita ser renovada para deshacerse
-                    // del 'Loading...' y todo en el hilo principal.
-                    DispatchQueue.main.async {
-                        self.hasSearched = false
-                        self.isLoading = false
-                        self.tableView.reloadData()
-                        self.showNetWorkError("Whoops", "There was an error reading from the iTunes Store. Please try again.")
-                    }
-                    
-                } else {
-                    print("Success! \(response!)")
-                }
-            })
-            
-            // 5.- Una vez creada la tarea de datos, llamar al método 'resume()' para inicializar el proceso,
-            // enviando una solicitud al server. Todo sucede en un subproceso en segundo plano, por lo que la
-            // aplicación es liberada inmediatamente para continuar (URLSession es asíncrona).
-            dataTask?.resume()
-        }
+        // Hacemos que la nueva clase creada "Search" haga el trabajo de búsqueda que se hacía antes aquí.
+        search.performSearch(for: searchBar.text!, category: segmentControl.selectedSegmentIndex, completion: { success in
+            if !success {
+                self.showNetWorkError("Error NetWork", "A connection error occurred when accessing our servers")
+            }
+            // Recargamos la tabla.
+            self.tableView.reloadData()
+        })
+        
+        // Recargamos la tabla y ocultamos el teclado.
+        tableView.reloadData()
+        searchBar.resignFirstResponder()
+
     }
     
     // Función para ajustar la posición de la 'searchBar' y que quede por encima de la tabla y
@@ -514,29 +280,29 @@ extension SearchViewController: UITableViewDataSource {
     // Número de filas que tendrá la tabla.
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         
-        if isLoading {
+        if search.isLoading {
             
             // Devolvemos uno porque necesitamos una fila para mostrar la celda de 'Cargando...'
-            return 1
+            return 1 // Loanding...
             
             // Si no hubo aún una búsqueda no devuelve ninguna fila.
-        } else if !hasSearched {
-            return 0
+        } else if !search.hasSearched {
+            return 0 // Not searched yet
             
             // Si no hay resultados, devuelve una fila.
-        } else if searchResults.count == 0 {
-            return 1
+        } else if search.searchResults.count == 0 {
+            return 1 // Nothing Found
             
             // Si hay resultados, devuleve tantas filas como contenga el array que almacena los datos.
         } else {
-            return searchResults.count
+            return search.searchResults.count
         }
     }
     
     // Configura la celda a mostrar para el indexpath dado.
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         
-        if isLoading {
+        if search.isLoading {
             let cell = tableView.dequeueReusableCell(withIdentifier: TableViewCellIdentifiers.loadingCell, for: indexPath)
             let spinner = cell.viewWithTag(100) as! UIActivityIndicatorView
             spinner.startAnimating()
@@ -544,7 +310,7 @@ extension SearchViewController: UITableViewDataSource {
         }
         
         // Si no hay resultados...
-        if searchResults.count == 0 {
+        if search.searchResults.count == 0 {
             // Devuleve la celda con el identificador de la celda "nada encontrado".
             return tableView.dequeueReusableCell(withIdentifier: TableViewCellIdentifiers.nothingFoundCell, for: indexPath)
             
@@ -555,7 +321,7 @@ extension SearchViewController: UITableViewDataSource {
             
             // Almacenamos el resultado para el "indexpath" de la fila/row correspondiente, y
             // asignamos los valores que tendrán las views de la celda, y devolvemos la celda.
-            let searchResult = searchResults[indexPath.row]
+            let searchResult = search.searchResults[indexPath.row]
             
             // Llamamos al método configure de la clase SearchResultCell.
             cell.configure(for: searchResult)
@@ -586,7 +352,7 @@ extension SearchViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, willSelectRowAt indexPath: IndexPath) -> IndexPath? {
         
         // Si no hay resultados, devuelve nil, de lo contrario, devuelve el nuevo indexpath seleccionado.
-        if searchResults.count == 0 || isLoading {
+        if search.searchResults.count == 0 || search.isLoading {
             return nil
         } else {
             return indexPath
